# be/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "RobotPet API",
        "description": "Flask API for RobotPet",
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000",
    "schemes": [
        "http"
    ],
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "JWT 토큰을 포함한 Authorization 헤더"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "회원가입",
                "description": "새 사용자로 등록 ",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "Password@123"
                                },
                                "confirm_password": {
                                    "type": "string",
                                    "example": "Password@123"
                                }
                            },
                            "required": [
                                "email",
                                "password",
                                "confirm_password"
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "회원가입 요청이 완료되었습니다. 이메일을 확인하세요.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "회원가입 요청이 완료되었습니다. 이메일을 확인하세요."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청, 비밀번호 확인 불일치 등"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/auth/verify-email": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "이메일 인증 (6자리 코드)",
                "description": "사용자가 이메일 인증을 위해 받은 6자리 코드를 입력하여 인증 완료",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                },
                                "code": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "required": [
                                "email",
                                "code"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "이메일 인증 성공",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "이메일 인증이 완료되었습니다."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 또는 인증 코드 불일치"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "로그인",
                "description": "이메일과 비밀번호를 사용한 로그인",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "Password@123"
                                },
                                "remember_me": {
                                    "type": "boolean",
                                    "example": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "로그인 성공 및 액세스 토큰 반환",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "access_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1..."
                                },
                                "refresh_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1..."
                                },
                                "redirect_url": {
                                    "type": "string",
                                    "example": "http://localhost:3000/home"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청"
                    },
                    "401": {
                        "description": "인증 실패"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "로그아웃",
                "description": "사용자 로그아웃 및 액세스 토큰 무효화",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "access_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1..."
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "로그아웃 성공"
                    },
                    "400": {
                        "description": "잘못된 요청"
                    },
                    "401": {
                        "description": "유효하지 않은 토큰"
                    }
                }
            }
        },
...
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "비밀번호 재설정",
                "description": "토큰을 이용해 비밀번호 변경",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                },
                                "new_password": {
                                    "type": "string",
                                    "example": "NewPassword123!"
                                },
                                "confirm_password": {
                                    "type": "string",
                                    "example": "NewPassword123!"
                                }
                            },
                            "required": [
                                "token",
                                "email",
                                "new_password",
                                "confirm_password"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "비밀번호 재설정 성공"
                    },
                    "400": {
                        "description": "잘못된 요청 또는 토큰 불일치"
                    }
                }
            },
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "비밀번호 재설정 페이지 로드",
                "description": "사용자가 비밀번호를 재설정할 때 토큰을 검증하여 해당 페이지를 로드",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "description": "비밀번호 재설정 토큰"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "토큰 검증 성공 및 비밀번호 재설정 페이지 로드",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "비밀번호 재설정 페이지를 로드합니다."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 토큰 또는 만료된 토큰"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
...
        "/auth/verify-email-request": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "이메일 인증 요청",
                "description": "사용자가 이메일을 입력하고 인증 요청 시 입력한 이메일로 인증 코드 전송",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                }
                            },
                            "required": [
                                "email"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "이메일 인증 코드 전송 완료",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "인증 코드가 이메일로 전송되었습니다."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (이메일 누락 또는 이미 가입된 이메일)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "이미 가입된 이메일입니다."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/auth/resend-verification-code": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "이메일 인증 코드 재전송",
                "description": "이미 인증 요청한 이메일에 인증 코드를 다시 전송",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "test1@test.com"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "새로운 인증 코드가 이메일로 전송됨"
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "액세스 토큰 갱신",
                "description": "리프레시 토큰을 사용해 새 액세스 토큰 발급",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "refresh_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1..."
                                }
                            },
                            "required": [
                                "refresh_token"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "새로운 액세스 토큰 발급",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "access_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1..."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (토큰 없음 또는 만료됨)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Refresh token has expired"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
...
        "/auth/verify-email-status": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "이메일 인증 상태 확인",
                "description": "이메일 인증했는지 상태 조회",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "example": "test1@test.com",
                        "description": "확인할 이메일 주소"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "이메일 인증 여부 조회",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "verified": {
                                    "type": "boolean",
                                    "example": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "이메일을 입력하지 않은 경우",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "이메일을 입력해야 합니다."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
...
    }
}